+++ lib/matplotlib/tests/baseline_images/test_backend_pdf/truetype-conversion.pdf
