+++ lib/matplotlib/tests/test_backend_pdf.py
@@ -434,3 +434,15 @@ def test_otf_font_smoke(family_name, file_name):
     fig = plt.figure()
     fig.text(0.15, 0.475, "Привет мир!")
     fig.savefig(io.BytesIO(), format="pdf")
+
+
+@image_comparison(["truetype-conversion.pdf"])
+# mpltest.ttf does not have "l"/"p" glyphs so we get a warning when trying to
+# get the font extents.
+def test_truetype_conversion(recwarn):
+    mpl.rcParams['pdf.fonttype'] = 3
+    fig, ax = plt.subplots()
+    ax.text(0, 0, "ABCDE",
+            font=Path(__file__).with_name("mpltest.ttf"), fontsize=80)
+    ax.set_xticks([])
+    ax.set_yticks([])
